Fix colours, background art, looping death sound and HUD overlap
Colours. The PPA's rgb_swap does a full byte reversal, not the R/B swap
its "ARGB becomes BGRA" wording suggests, and without it a BS pixel
already lines up with the panel's RGB888. Turning it on transposed reds
and blues, so PPA_FLIP_RGB_SWAP is now false.

Background art. Enabling it in CMakeLists did nothing: globals.h defined
DISABLE_BACKGROUND_ART as well, and either one turns it off. With both
cleared it is genuinely on for the first time, which is also why dying
showed a black screen -- showPicture drew a black rect instead of
livelost.jpg.

That exposed the memory cost. initBackground copied the artwork into a
surface padded by one screen in each direction so a single blit could run
off the edge, then made an immediate duplicate of that (SDL_DisplayFormat
is a plain copy here) while the decoded original was still alive.
LostPixels' largest background is 3328x952, so that peaked near 57 MB on a
32 MB device. Keep the artwork as loaded and wrap at draw time instead: up
to four blits on the frames that straddle an edge, one full-screen blit
otherwise -- still the shape the PPA path picks up. Same for the parallax
layer, and drop the same duplicate-and-discard in initGui.

Death sound looping forever: the voice cursor was 16.16 fixed point in a
uint32, which wraps after 65536 frames -- 1.49 s at 44100 Hz -- so any
longer effect restarted instead of ending. Widened to 64-bit.

Enter did double duty: it is what JOYSTICK_ACTION maps to on this
keyboard, but it also toggled the music, so every use of the action --
the antigrav field in LostPixels level 8, say -- flipped the music too.
Music moved to M. The action itself was level-triggered, so holding the
key re-fired it every frame; it now fires once, on release, like pause.

The end screens (highscore wait, intro, attract mode) read the held
button state, so a key still down from the previous screen dismissed each
one the instant it appeared -- pressing space on the death screen ran
straight through to a new game. They read release edges now, and flush on
entry.

HUD overlap: the status line sat in fixed columns spaced for the original
TTF face. The Hershey vector font is much wider -- "Pixels: 123 (456)"
alone overruns the 170px it was given -- so the fields ran into each other
and the time bar ran through its own label. Lay both rows out from
measured text widths via the new fontHandlerTextWidth().

+++ main/pal/pal_audio.cpp
@@ -76,7 +76,10 @@ typedef struct {
     const int16_t* samples;
     size_t         num_samples;
     int            channels;
-    uint32_t       pos_q16;     // read cursor in frames, 16.16 fixed point
+    // 16.16 fixed point, and 64-bit on purpose: in 32 bits the cursor wraps
+    // after 65536 frames -- 1.49 s at 44100 Hz -- and the voice restarts
+    // instead of ending, which left the death sound looping forever.
+    uint64_t       pos_q16;     // read cursor in frames, 16.16 fixed point
     uint32_t       step_q16;    // frames advanced per output frame
     bool           active;
 } Voice;
@@ -180,7 +183,7 @@ static void voice_start(const PCM_FX_Entry* e) {
     // All voices busy: steal the one furthest through its sample, which is
     // the one closest to finishing anyway.
     if (!v) {
-        uint32_t best = 0;
+        uint64_t best = 0;
         for (int i = 0; i < MAX_VOICES; i++) {
             if (s_voices[i].pos_q16 >= best) { best = s_voices[i].pos_q16; v = &s_voices[i]; }
         }
@@ -215,7 +218,7 @@ static void mix_voices(int32_t* accum, int frames) {
 
         size_t total_frames = v->num_samples / (size_t)v->channels;
         for (int f = 0; f < frames; f++) {
-            size_t src_frame = v->pos_q16 >> 16;
+            size_t src_frame = (size_t)(v->pos_q16 >> 16);
             if (src_frame >= total_frames) { v->active = false; break; }
             const int16_t* sp = v->samples + src_frame * (size_t)v->channels;
             int32_t l = sp[0];
